requirements.txt.txt
pandas
numpy 
matplotli 
statsmodels
yfinance